Reduce the implementation requirements for `delete-if`
In implementations derived from KCL, such as GCL, ECL, and Clasp, it
is possible for `delete` to call its `test` multiple times on the same
element. No evidence has been found to suggest that this behavior is
inconsistent with the standard. Therefore, additional checks need to
be added to ensure that tweens and timelines can be properly freed and
removed through a single traversal only.